# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login with email & password",
                "description": "This login api customer can login with email & password",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "mdrabiulhasan.me@gmail.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": 200,
                                    "success": true,
                                    "message": "Login successfully",
                                    "data": {
                                        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8xMjcuMC4wLjE6ODAwMFwvYXBpXC9hdXRoXC9sb2dpbiIsImlhdCI6MTY1MzU1NDE0NiwiZXhwIjoxNjUzNTU3NzQ2LCJuYmYiOjE2NTM1NTQxNDYsImp0aSI6IlhseGpwWERmYnBrZzJQMmYiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.763UrUujMp_n6GFC5yR1ahsdkzq_6JF53xJSv8A_e2g",
                                        "token_type": "bearer",
                                        "name": "Rabiul Hasan",
                                        "expires_in": 3600
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Email & Password Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": 400,
                                    "success": false,
                                    "message": "Your email or password was incorrect"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 403
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "New Consumer Registration",
                "description": "This register api can add new consumer in our application",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "password_confirmation": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Rabiul Hasan",
                                    "email": "mdrabiulhasan.me@gmail.com",
                                    "password": "123456",
                                    "password_confirmation": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully Register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "User successfully registered",
                                    "user": {
                                        "name": "Rabiul Hasan",
                                        "email": "mdrabiulhasan1.me@gmail.com",
                                        "updated_at": "2022-05-26T09:26:17.000000Z",
                                        "created_at": "2022-05-26T09:26:17.000000Z",
                                        "id": 2
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/user-profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticated User Profile",
                "description": "This api fetch authentication user profile",
                "operationId": "00e8543ad81adb67d8bb6c4add8ff4af",
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "name": "Rabiul Hasan",
                                    "email": "mdrabiulhasan.me@gmail.com",
                                    "email_verified_at": null,
                                    "created_at": "2022-05-26T03:49:08.000000Z",
                                    "updated_at": "2022-05-26T03:49:08.000000Z"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt_token": []
                    }
                ]
            }
        },
        "/auth/name/{name}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Url Parameter Pass",
                "description": "This api fetch url parameter pass",
                "operationId": "96fc4144c87ffc149e21aa528b614316",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "status": 200,
                                    "success": true,
                                    "message": "Your name is Rabiul-Hasan"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Login/Registration Route Here"
        }
    ],
    "components": {
        "securitySchemes": {
            "jwt_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}